Enhance code
