+++ web/src/pages/role-perimission/components/role-from/props.ts
@@ -1,0 +1,15 @@
+export type DepartmentDto = {
+  [key: string]: string | boolean | undefined;
+  name: string;
+  id: string;
+  isSelected: boolean;
+  isExpand?: boolean;
+  indeterminate?: boolean;
+  parentId?: string;
+  isHide?: boolean;
+};
+
+export type AllDepartmentData = {
+  pullCrowdData: DepartmentDto[];
+  notificationData: DepartmentDto[];
+};
